[refactor] Move usRoutes to middleware folder
